Created two scenes for the demo, added script called changeScenes to change between scenes
Organized project directory a little by placing all scenes in a scene folder. Created a prefab folder for the demo. Imported custom made house for use in the demo. Created a scripts folder for scripts. Added a changeScenes script with a template code that can be used to change between scenes. Created two scenes, one with a piece of cheese on the table, and one with an apple on the table. Placed camera at table.